refactor(components): [popover] switch to script-setup syntax (#8273)
