+++ packages/components/popover/__tests__/directive.test.ts
@@ -2,7 +2,7 @@ import { nextTick, ref } from 'vue'
 import { afterEach, describe, expect, test } from 'vitest'
 import makeMount from '@element-plus/test-utils/make-mount'
 import { rAF } from '@element-plus/test-utils/tick'
-import Popover from '../src/index.vue'
+import Popover from '../src/popover.vue'
 import PopoverDirective, { VPopover } from '../src/directive'
 
 const AXIOM = 'Rem is the best girl'
