+++ packages/components/popover/__tests__/popover.test.ts
@@ -4,7 +4,7 @@ import { POPPER_CONTAINER_SELECTOR, useZIndex } from '@element-plus/hooks'
 import makeMount from '@element-plus/test-utils/make-mount'
 import { rAF } from '@element-plus/test-utils/tick'
 import { ElPopperTrigger } from '@element-plus/components/popper'
-import Popover from '../src/index.vue'
+import Popover from '../src/popover.vue'
 
 const AXIOM = 'Rem is the best girl'
